Remove HotPreview annotations and update sonar exclusions
This commit removes the HotPreview annotations from various composable functions to clean up the codebase and improve maintainability. Additionally, the sonar exclusions have been updated to exclude the debugMain source set, ensuring that unnecessary files are not included in the code quality analysis.
