+++ composeApp/build.gradle.kts
@@ -66,6 +66,10 @@ kotlin {
   @OptIn(ExperimentalKotlinGradlePluginApi::class)
   applyDefaultHierarchyTemplate {
     common {
+      group("debug") {
+        withAndroidTarget()
+        withJvm()
+      }
       group("nonJs") {
         withAndroidTarget()
         withJvm()
@@ -113,9 +117,14 @@ kotlin {
       implementation(libs.koin.compose)
       implementation(libs.koin.compose.viewmodel)
       implementation(libs.koin.compose.viewmodel.navigation)
+    }
 
-      // Hot Preview
-      implementation(libs.hotpreview)
+    // Debug source set configuration
+    val debugMain by getting {
+      dependencies {
+        // Hot Preview
+        implementation(libs.hotpreview)
+      }
     }
 
     androidMain.dependencies {
