+++ composeApp/src/commonMain/kotlin/proj/memorchess/axl/ui/components/board/BoardGrid.kt
@@ -6,21 +6,14 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.rememberCoroutineScope
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import de.drick.compose.hotpreview.HotPreview
-import de.drick.compose.hotpreview.HotPreviewParameter
-import de.drick.compose.hotpreview.HotPreviewParameterProvider
 import kotlinx.coroutines.launch
 import memorchess.composeapp.generated.resources.Res
 import memorchess.composeapp.generated.resources.description_board_tile
 import org.jetbrains.compose.resources.stringResource
-import proj.memorchess.axl.core.config.CHESS_BOARD_COLOR_SETTING
-import proj.memorchess.axl.core.interactions.LinesExplorer
-import proj.memorchess.axl.ui.theme.ChessBoardColorScheme
 
 @Composable
 fun BoardGrid(state: BoardGridState, modifier: Modifier = Modifier) {
@@ -53,32 +46,4 @@ fun BoardGrid(state: BoardGridState, modifier: Modifier = Modifier) {
       }
     }
   }
-}
-
-@HotPreview(
-  density = 2.625f,
-  widthDp = 411,
-  heightDp = 891,
-  statusBar = true,
-  navigationBar = de.drick.compose.hotpreview.NavigationBarMode.GestureBottom,
-  displayCutout = de.drick.compose.hotpreview.DisplayCutoutMode.CameraTop,
-)
-@Composable
-private fun BoardGridPreview(
-  @HotPreviewParameter(BoardGridPreviewProvider::class) params: ChessBoardColorScheme
-) {
-  CHESS_BOARD_COLOR_SETTING.setValue(params)
-  Column(modifier = Modifier.fillMaxSize()) {
-    Text(params.displayName, modifier = Modifier.weight(1f))
-    Box(modifier = Modifier.weight(8f)) {
-      BoardGrid(BoardGridState(false, LinesExplorer()), Modifier.fillMaxSize())
-    }
-  }
-}
-
-class BoardGridPreviewProvider : HotPreviewParameterProvider<ChessBoardColorScheme> {
-  override val values: Sequence<ChessBoardColorScheme>
-    get() {
-      return ChessBoardColorScheme.entries.asSequence()
-    }
 }
