+++ composeApp/src/commonMain/kotlin/proj/memorchess/axl/ui/components/board/Tile.kt
@@ -2,20 +2,12 @@ package proj.memorchess.axl.ui.components.board
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.Text
 import androidx.compose.runtime.*
 import androidx.compose.ui.Modifier
-import de.drick.compose.hotpreview.HotPreview
-import de.drick.compose.hotpreview.HotPreviewParameter
-import de.drick.compose.hotpreview.HotPreviewParameterProvider
 import proj.memorchess.axl.core.config.CHESS_BOARD_COLOR_SETTING
 import proj.memorchess.axl.core.engine.board.ITile
 import proj.memorchess.axl.core.engine.board.ITile.TileColor
-import proj.memorchess.axl.core.engine.board.Tile
-import proj.memorchess.axl.ui.theme.ChessBoardColorScheme
 
 @Composable
 fun Tile(tile: ITile) {
@@ -28,26 +20,4 @@ fun Tile(tile: ITile) {
         )
         .fillMaxSize()
   )
-}
-
-@HotPreview(density = 1.0f, widthDp = 200, heightDp = 100, captionBar = false)
-@Composable
-private fun TilePreview(
-  @HotPreviewParameter(TilePreviewProvider::class) params: ChessBoardColorScheme
-) {
-  CHESS_BOARD_COLOR_SETTING.setValue(params)
-  Column(modifier = Modifier.fillMaxSize()) {
-    Text(params.displayName, modifier = Modifier.weight(1f))
-    Row(modifier = Modifier.weight(5f)) {
-      Box(modifier = Modifier.weight(1f)) { Tile(Tile(0, 0)) }
-      Box(modifier = Modifier.weight(1f)) { Tile(Tile(0, 1)) }
-    }
-  }
-}
-
-class TilePreviewProvider : HotPreviewParameterProvider<ChessBoardColorScheme> {
-  override val values: Sequence<ChessBoardColorScheme>
-    get() {
-      return ChessBoardColorScheme.entries.asSequence()
-    }
 }
