+++ composeApp/src/commonMain/kotlin/proj/memorchess/axl/ui/layout/explore/PortraitExploreLayout.kt
@@ -8,13 +8,11 @@ import androidx.compose.foundation.layout.padding
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import de.drick.compose.hotpreview.HotPreview
 import proj.memorchess.axl.ui.components.explore.ExploreActionButtons
 import proj.memorchess.axl.ui.components.explore.ExploreBoardSection
 import proj.memorchess.axl.ui.components.explore.ExploreHeader
 import proj.memorchess.axl.ui.components.explore.ExploreNextMovesSection
 import proj.memorchess.axl.ui.components.explore.ExploreStateIndicators
-import proj.memorchess.axl.ui.util.previewExploreLayoutContent
 
 @Composable
 fun PortraitExploreLayout(modifier: Modifier = Modifier, content: ExploreLayoutContent) {
@@ -47,10 +45,4 @@ fun PortraitExploreLayout(modifier: Modifier = Modifier, content: ExploreLayoutC
       ExploreActionButtons(saveButton = content.saveButton, deleteButton = content.deleteButton)
     }
   }
-}
-
-@HotPreview(widthDp = 411, heightDp = 891, density = 2.625f)
-@Composable
-private fun PortraitExploreLayoutPreview() {
-  PortraitExploreLayout(content = previewExploreLayoutContent)
 }
