+++ composeApp/src/debugMain/kotlin/proj/memorchess/axl/preview/layouts/ExploreLayoutContent.kt
@@ -1,10 +1,14 @@
-package proj.memorchess.axl.ui.util
+package proj.memorchess.axl.preview.layouts
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.material3.*
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import compose.icons.FeatherIcons
@@ -19,7 +23,7 @@ import proj.memorchess.axl.ui.components.board.StateIndicator
 import proj.memorchess.axl.ui.components.buttons.ControlButton
 import proj.memorchess.axl.ui.layout.explore.ExploreLayoutContent
 
-val previewExploreLayoutContent =
+internal val previewExploreLayoutContent =
   ExploreLayoutContent(
     resetButton = { ControlButton.RESET.render(it) {} },
     reverseButton = { ControlButton.REVERSE.render(it) {} },
