+++ composeApp/src/debugMain/kotlin/proj/memorchess/axl/preview/training/DaysInAdvanceCardPreview.kt
@@ -1,0 +1,11 @@
+package proj.memorchess.axl.preview.training
+
+import androidx.compose.runtime.Composable
+import de.drick.compose.hotpreview.HotPreview
+import proj.memorchess.axl.ui.components.training.DaysInAdvanceCard
+
+@Composable
+@HotPreview
+private fun DaysInAdvanceCardPreview() {
+  DaysInAdvanceCard(days = 5)
+}
